refactoring tests files
