+++ server/tests/message.test.js
@@ -1,6 +1,6 @@
 var expect = require("expect");
 
-var {generateMessage, generateLocationMessage} = require("./message");
+var {generateMessage, generateLocationMessage} = require("../utils/message");
 
 describe("generateMessage", () => {
     it("should generate a correct message", () => {
